Support for UserDefaultsValue instances to be observed by multiple observers at once
