+++ nightguard/UserDefaultsValueGroups.swift
@@ -73,16 +73,3 @@ extension UserDefaultsValueGroups {
         static let alarm = "alarm"
     }
 }
-
-class ObservationToken {
-    
-    private let cancellationClosure: () -> Void
-    
-    init(cancellationClosure: @escaping () -> Void) {
-        self.cancellationClosure = cancellationClosure
-    }
-    
-    func cancel() {
-        cancellationClosure()
-    }
-}
